Move the batch API into its own file
Still exported in the blobs module though.
